pyproject.toml
[project]
name = "langsmith-otel-redact"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "certifi==2026.2.25",
    "charset-normalizer==3.4.4",
    "distro==1.9.0",
    "googleapis-common-protos==1.72.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.11",
    "importlib-metadata==8.7.1",
    "jiter==0.13.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==1.2.10",
    "langchain-core==1.2.16",
    "langchain-openai==1.1.10",
    "langgraph==1.0.9",
    "langgraph-checkpoint==4.0.0",
    "langgraph-prebuilt==1.0.8",
    "langgraph-sdk==0.3.9",
    "langsmith==0.7.7",
    "openai==2.24.0",
    "opentelemetry-api==1.39.1",
    "opentelemetry-exporter-otlp-proto-common==1.39.1",
    "opentelemetry-exporter-otlp-proto-http==1.39.1",
    "opentelemetry-proto==1.39.1",
    "opentelemetry-sdk==1.39.1",
    "opentelemetry-semantic-conventions==0.60b1",
    "orjson==3.11.7",
    "ormsgpack==1.12.2",
    "packaging==26.0",
    "protobuf==6.33.5",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "python-dotenv==1.2.1",
    "pyyaml==6.0.3",
    "regex==2026.2.19",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "sniffio==1.3.1",
    "tenacity==9.1.4",
    "tiktoken==0.12.0",
    "tqdm==4.67.3",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.3",
    "uuid-utils==0.14.1",
    "xxhash==3.6.0",
    "zipp==3.23.0",
    "zstandard==0.25.0",
]
